Add about page - project suspended
Decided to suspend the development of the Gatsby version of the web,
due to the high complexity involving i18n and content translation.

It is found that other static site generator, such as Jekyll, has a
better approach to content translation than Gatsby, especially in
the context of static website.

Gatsby i18n relies on the locale files to translate each bit of text to
different language independently, which makes it a very good way to
translate a web application into various languages.
However, this appoarch is not suitable for translation of content-rich
website such as blogs. Mainly because most of the time the text are not
translated as a direct one-to-one translation.
This means that different languages could have different about of text,
which makes the programming very complicated.
The only manageble way is to store the entire text -- format included --
in the locale file. This violates the principle of context-style
saparation and leads to a very annoying state of content writing.

Comparing this approach to Jekyll's "different content file for a
different language", where one simply makes another file with the
same name, with language code suffix, containing the entire translation.
While this approach makes a lot of redundancy and not suitable for
web application translation, it seems to be the best approach for
translating content-rich website such as a blog.

Taking this issue into consideration, the Gatsby project is suspended.
The website will not revert to Jekyll version.